fix(cli): handle missing target artifact dir
When creating a new artifact (e.g. Model or Repository) while the target
directory does not exist yet (e.g. `src/models` or `src/repositories`),
we used to abort CLI with a ENOENT error.

This commit is fixing our CLI to treat the missing directory as if it
was empty and continue the operation if possible.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>